// packages/manager/apps/okms/src/common/types/location.type.ts
type CardinalPoint =
  | 'CENTRAL'
  | 'EAST'
  | 'NORTH'
  | 'NORTHEAST'
  | 'NORTHWEST'
  | 'SOUTH'
  | 'SOUTHEAST'
  | 'SOUTHWEST'
  | 'WEST';

type SpecificType = 'BACKUP' | 'LZ' | 'SNC' | 'STANDARD';

export type LocationType = 'LOCAL-ZONE' | 'REGION-1-AZ' | 'REGION-3-AZ';

export type GeographyCode = string;

export type Location = {
  availabilityZones: string[];
  cardinalPoint: CardinalPoint;
  cityCode: string;
  cityLatitude: number;
  cityLongitude: number;
  cityName: string;
  code: string;
  countryCode: string;
  countryName: string;
  geographyCode: GeographyCode;
  geographyName: string;
  location: string;
  name: string;
  openingYear: number;
  specificType: SpecificType;
  type: LocationType;
};